brought the entire blackjack program into one file. feelsgoodman
